move asts tests from serial to perodical job
